make team mates position a double

+++ module/network/RobotCommunication/src/RobotCommunication.cpp
@@ -137,7 +137,7 @@ namespace module::network {
 
             for (auto& mate : teammates.teammates) {  // Loop through each teammate.
                 if (mate.id == id) {  // If the ID of a teammate is found to be the same, update the position.
-                    mate.rRFf = robocup.current_pose.position;  // Update position
+                    mate.rRFf = robocup.current_pose.position.cast<double>();  // Update position
                     found     = true;
                     break;  // Break out of the loop.
                 }
@@ -148,7 +148,7 @@ namespace module::network {
                 TeamMate mate;
                 // Set the variables
                 mate.id   = id;
-                mate.rRFf = robocup.current_pose.position;
+                mate.rRFf = robocup.current_pose.position.cast<double>();
                 // Add it to the teammates list
                 teammates.teammates.emplace_back(mate);
             }
